gestion des erreurs custom

+++ src/controllers/freezertypeCtrl.js
@@ -1,49 +1,53 @@
-// import des modules nécessaires
 const DB = require('../config/db.config');
 const FreezerType = DB.FreezerType;
+const {
+	RequestError,
+	FreezerTypeError,
+	ForbiddenError,
+} = require('../errors/customErrors');
 
-exports.getAllFreezerTypes = async (req, res) => {
+exports.getAllFreezerTypes = async (req, res, next) => {
 	try {
 		let freezerType = await FreezerType.findAll();
 		return res.json({ data: freezerType });
 	} catch (error) {
-		return res.status(500).json({ message: 'Database error', error: error });
+		next(error);
 	}
 };
 
-exports.getFreezerTypeById = async (req, res) => {
-	let id = parseInt(req.params.id);
-
-	if (!id) {
-		return res.status(400).json({ message: 'Missing parameters' });
-	}
-
+exports.getFreezerTypeById = async (req, res, next) => {
 	try {
+		let id = parseInt(req.params.id);
+
+		if (!id) {
+			throw new RequestError('Missing parameter');
+		}
+
 		let freezerType = await FreezerType.findOne({
 			where: { id: id },
 			raw: true,
 		});
 
 		if (freezerType === null) {
-			return res.status(404).json({ message: 'Freezer Type not found' });
+			throw new FreezerTypeError('Freezer type not found', 3);
 		}
 
 		return res.json({ data: freezerType });
 	} catch (error) {
-		return res.status(500).json({ message: 'Database error', error: error });
+		next(error);
 	}
 };
 
-exports.createFreezerType = async (req, res) => {
-	if (!req.body.name) {
-		return res.status(400).json({ message: 'Missing data' });
-	}
-
-	if (!req.isAdmin) {
-		return res.status(403).json({ message: 'Forbidden' });
-	}
-
+exports.createFreezerType = async (req, res, next) => {
 	try {
+		if (!req.body.name) {
+			throw new RequestError('Missing data');
+		}
+
+		if (!req.isAdmin) {
+			throw new ForbiddenError('Forbidden');
+		}
+
 		let freezerType = await FreezerType.findOne({
 			where: { name: req.body.name.trim().toLowerCase() },
 			raw: true,
@@ -53,9 +57,10 @@ exports.createFreezerType = async (req, res) => {
 			freezerType !== null &&
 			freezerType.name.toLowerCase() == req.body.name.trim().toLowerCase()
 		) {
-			return res.status(409).json({
-				message: `The freezer type '${req.body.name}' already exists`,
-			});
+			throw new FreezerTypeError(
+				`The freezer type '${req.body.name}' already exists`,
+				null
+			);
 		}
 
 		// On nettoie et on formate le nom
@@ -65,35 +70,36 @@ exports.createFreezerType = async (req, res) => {
 
 		return res.json({ message: 'Freezer Type created', data: newFreezerType });
 	} catch (error) {
-		return res.status(500).json({ message: 'Database error', error: error });
+		next(error);
 	}
 };
 
-exports.updateFreezerType = async (req, res) => {
-	let id = parseInt(req.params.id);
-
-	if (!id) {
-		return res.status(400).json({ message: 'Missing parameters' });
-	}
-
-	if (!req.isAdmin) {
-		return res.status(403).json({ message: 'Forbidden' });
-	}
-
+exports.updateFreezerType = async (req, res, next) => {
 	try {
+		let id = parseInt(req.params.id);
+
+		if (!id) {
+			throw new RequestError('Missing parameter');
+		}
+
+		if (!req.isAdmin) {
+			throw new ForbiddenError('Forbidden');
+		}
+
 		let freezerType = await FreezerType.findOne({
 			where: { id: id },
 			raw: true,
 		});
 
 		if (freezerType === null) {
-			return res.status(404).json({ message: 'Freezer Type not found' });
+			throw new FreezerTypeError('Freezer type not found', 3);
 		} else if (
 			freezerType.name.toLowerCase() == req.body.name.trim().toLowerCase()
 		) {
-			return res.status(409).json({
-				message: `The freezer type '${req.body.name}' already exists`,
-			});
+			throw new FreezerTypeError(
+				`The freezer type '${req.body.name}' already exists`,
+				null
+			);
 		}
 
 		// On nettoie et on formate le nom
@@ -103,22 +109,30 @@ exports.updateFreezerType = async (req, res) => {
 
 		return res.json({ message: 'Freezer Type updated', data: newFt });
 	} catch (error) {
-		return res.status(500).json({ message: 'Database error', error: error });
+		next(error);
 	}
 };
 
-exports.deleteFreezerType = async (req, res) => {
-	let id = parseInt(req.params.id);
-
-	if (!id) {
-		return res.status(400).json({ message: 'Missing parameters' });
-	}
-
-	if (!req.isAdmin) {
-		return res.status(403).json({ message: 'Forbidden' });
-	}
-
+exports.deleteFreezerType = async (req, res, next) => {
 	try {
+		let id = parseInt(req.params.id);
+
+		if (!id) {
+			throw new RequestError('Missing parameter');
+		}
+
+		if (!req.isAdmin) {
+			throw new ForbiddenError('Forbidden');
+		}
+
+		let freezerType = await FreezerType.findOne({
+			where: { id: id },
+		});
+
+		if (freezerType === null) {
+			throw new FreezerTypeError('Freezer type not found', 3);
+		}
+
 		// Supression du freezer type
 		await FreezerType.destroy({ where: { id: id }, force: true });
 
@@ -128,41 +142,54 @@ exports.deleteFreezerType = async (req, res) => {
 	}
 };
 
-exports.trashFreezerType = async (req, res) => {
-	let id = parseInt(req.params.id);
-
-	if (!id) {
-		return res.status(400).json({ message: 'Missing parameters' });
-	}
-
-	if (!req.isAdmin) {
-		return res.status(403).json({ message: 'Forbidden' });
-	}
-
+exports.trashFreezerType = async (req, res, next) => {
 	try {
+		let id = parseInt(req.params.id);
+
+		if (!id) {
+			throw new RequestError('Missing parameters');
+		}
+
+		if (!req.isAdmin) {
+			throw new ForbiddenError('Forbidden');
+		}
+
+		let freezerType = await FreezerType.findOne({
+			where: { id: id },
+		});
+
+		if (freezerType === null) {
+			throw new FreezerTypeError('Freezer type not found', 3);
+		}
 		// Soft delete
 		await reezerType.destroy({ where: { id: id } });
 		return res.status(204).json({ message: 'Freezer type soft deleted' });
 	} catch (error) {
-		return res.status(500).json({ message: 'Database error', error: error });
+		next(error);
 	}
 };
 
-exports.untrashFreezerType = async (req, res) => {
-	let id = parseInt(req.params.id);
-
-	if (!id) {
-		return res.status(400).json({ message: 'Missing parameters' });
-	}
-
-	if (!req.isAdmin) {
-		return res.status(403).json({ message: 'Forbidden' });
-	}
-
+exports.untrashFreezerType = async (req, res, next) => {
 	try {
+		let id = parseInt(req.params.id);
+
+		if (!id) {
+			throw new RequestError('Missing parameters');
+		}
+
+		if (!req.isAdmin) {
+			throw new ForbiddenError('Forbidden');
+		}
+		let freezerType = await FreezerType.findOne({
+			where: { id: id },
+		});
+
+		if (freezerType === null) {
+			throw new FreezerTypeError('Freezer type not found', 3);
+		}
 		await FreezerType.restore({ where: { id: id } });
 		return res.status(204).json({ message: 'Freezer type restored' });
 	} catch (error) {
-		return res.status(500).json({ message: 'Database error', error: error });
+		next(error);
 	}
 };
